[WIP] refactor spec helpers, begin bd fixture deployments
